backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
sqlalchemy==2.0.43
alembic==1.16.4
psycopg[binary]==3.2.9
email-validator==2.2.0
